requirements.txt
torch
torchvision
opencv-python
mediapipe
numpy
pillow
matplotlib
scikit-learn
tqdm
